[#16273] - promoting abstract locator to support:
